+++ pymongo/pool.py
@@ -560,7 +560,7 @@ class SocketInfo(object):
                 'Must be connected to MongoDB 3.4+ to use a collation.')
 
         if session:
-            session._apply_to(spec, retryable_write, read_preference)
+            session._apply_to(spec, retryable_write, read_preference, self)
         self.send_cluster_time(spec, session, client)
         listeners = self.listeners if publish_events else None
         unacknowledged = write_concern and not write_concern.acknowledged